requirements.txt
Flask
groq
langchain_community
bs4
langchain-groq
flask-pydantic
